Remove installer
This is part of a series of changes for the `3.0` line of this library. This change, specifically, removes the installer, as we're no longer providing that as part of onboarding. Our new approach is to split off 3rd party integration into their own specific library. For example, our `ActiveRecord` integraiton will have `timber_active_record` and our `Rack` integration will have `timber_rack`. This provides users with an a le carte approach to structuring their logs as well as simple focused installation. Users can start with the core Timber library by simply installing the logger and proceed to install integrations as they get more comfortable. This makes adoption and installation less overwhelming, avoiding the need for an installer.